refactor: migrate OpenChoreo auth, permission and fetch API factories to core plugin
- Removed local implementations of OpenChoreo fetch, permission, and auth APIs from the portal app.
- Integrated OpenChoreo API references from the new core plugin, ensuring proper encapsulation and reuse.
- Updated tests and components to utilize the new API structure.
- Adjusted README files across plugins to reflect the dependency on the core OpenChoreo plugin.
