+++ plugins/openchoreo-ci/package.json
@@ -70,6 +70,7 @@
     "yaml": "^2.8.1"
   },
   "peerDependencies": {
+    "@openchoreo/backstage-plugin": "^1.2.0",
     "react": "^18.0.0",
     "react-router-dom": "^6.0.0"
   },
