+++ plugins/openchoreo-workflows/README.md
@@ -1,0 +1,10 @@
+# openchoreo-workflows
+
+Backstage frontend plugin for OpenChoreo generic workflow runs.
+
+## Requirements
+
+Requires [`@openchoreo/backstage-plugin`](../openchoreo) (the OpenChoreo
+core plugin) to be installed in the same Backstage app. The core plugin
+provides the fetch / permission / auth APIs this plugin uses to reach the
+OpenChoreo backend.
