+++ plugins/openchoreo-workflows/package.json
@@ -67,6 +67,7 @@
     "yaml": "^2.8.2"
   },
   "peerDependencies": {
+    "@openchoreo/backstage-plugin": "^1.2.0",
     "react": "^18.0.0",
     "react-router-dom": "^6.0.0"
   },
