+++ plugins/platform-engineer-core/package.json
@@ -67,6 +67,7 @@
     "react-use": "17.6.0"
   },
   "peerDependencies": {
+    "@openchoreo/backstage-plugin": "^1.2.0",
     "react": "^18.0.0",
     "react-dom": "^18.0.0",
     "react-router-dom": "^6.0.0"
